refactor: login to home page
